be/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
pandas==2.2.3
geopandas==1.0.1
shapely==2.1.0
pydantic==2.11.4
python-multipart==0.0.20
openai>=1.30.0
python-dotenv>=1.0.0
